requirements.txt
aiofiles==22.1.0
attrs==22.1.0
black==22.10.0
certifi==2022.9.24
click==8.1.3
colorama==0.4.5
edgedb==0.24.0
flake8==5.0.4
httptools==0.5.0
iniconfig==1.1.1
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
multidict==6.0.2
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
pytest==7.1.3
python-dotenv==0.21.0
sanic==22.9.0
sanic-routing==22.8.0
tomli==2.0.1
typing_extensions==4.4.0
websockets==10.3
